+++ toad/__init__.py
@@ -12,7 +12,7 @@ from .metrics import KS, KS_bucket, F1
 from .stats import quality, IV, VIF, WOE, entropy, entropy_cond, gini, gini_cond
 from .selection import select
 from .scorecard import ScoreCard
-from .utils import Progress
+from .utils import Progress, performance
 from .version import __version__
 
 VERSION = __version__